Integrated the neo4j garph visualization with markdown
